backend/requirements.txt
semantic-kernel==1.9.0
azure-ai-openai==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
aiohttp==3.9.1
openai==1.3.7
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
faker==20.1.0
seaborn>=0.12.0
matplotlib>=3.7.0
